# eyeguard/OpenAPI.yaml
...
openapi: 3.0.3
info:
  title: EyeGuard API
  version: 1.0.0
  description: |
    Software-only simulation REST API for the EyeGuard SOC training platform.
servers:
  - url: https://localhost:8000
    description: Local FastAPI dev server
paths:
...
  /api/v1/devices:
    get:
      summary: List devices
      operationId: listDevices
      responses:
        '200':
          description: Devices collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    post:
      summary: Create simulated device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          description: Duplicate IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/devices/{deviceId}:
    get:
      summary: Retrieve device detail
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update device metadata
      operationId: updateDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
  /api/v1/devices/{deviceId}/block:
    post:
      summary: Toggle automatic block
      operationId: blockDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device block toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
  /api/v1/alerts:
    get:
      summary: List alerts
      operationId: listAlerts
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [Low, Medium, High]
        - name: status
          in: query
          schema:
            type: string
            enum: [Open, Acknowledged, Resolved]
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
    post:
      summary: Create manual alert
      operationId: createAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertCreate'
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /api/v1/alerts/{alertId}:
    get:
      summary: Alert detail
      operationId: getAlert
      parameters:
        - $ref: '#/components/parameters/AlertId'
      responses:
        '200':
          description: Alert detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/alerts/{alertId}/status:
    post:
      summary: Update alert status
      operationId: updateAlertStatus
      parameters:
        - $ref: '#/components/parameters/AlertId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Open, Acknowledged, Resolved]
      responses:
        '200':
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
...
  /api/v1/reports/{reportId}:
    get:
      summary: Report detail
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetail'
  /api/v1/reports/{reportId}/export.csv:
    get:
      summary: Export report CSV
      operationId: exportReportCsv
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/v1/reports/{reportId}/export.pdf:
    get:
      summary: Export report PDF
      operationId: exportReportPdf
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /api/v1/simulation/devices:
    post:
      summary: Add simulated device to session
      operationId: addSimulationDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationDeviceCreate'
      responses:
        '201':
          description: Simulation device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationDevice'
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AlertId:
      name: alertId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ip_address:
          type: string
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
        status:
          type: string
        last_seen_at:
          type: string
          format: date-time
    DeviceCreate:
      type: object
      required: [ip_address, hostname, device_type, owner_role]
      properties:
        ip_address:
          type: string
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
    DeviceUpdate:
      type: object
      properties:
        hostname:
          type: string
        device_type:
          type: string
        owner_role:
          type: string
        traffic_gb:
          type: number
        status:
          type: string
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        detected_at:
          type: string
          format: date-time
        source_ip:
          type: string
        destination_ip:
          type: string
        category:
          type: string
        severity:
          type: string
        status:
          type: string
        rationale:
          type: string
        action_taken:
          type: string
    AlertCreate:
      type: object
      required: [source_ip, category, severity]
      properties:
        source_ip:
          type: string
        destination_ip:
          type: string
        category:
          type: string
        severity:
          type: string
        rationale:
          type: string
    AlertDetail:
      allOf:
        - $ref: '#/components/schemas/Alert'
        - type: object
          properties:
            events:
              type: array
              items:
                type: object
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alert_id:
          type: string
          format: uuid
        summary:
          type: string
        remediation_steps:
          type: string
        indicators:
          type: array
          items:
            type: string
    ReportDetail:
      allOf:
        - $ref: '#/components/schemas/Report'
    SimulationDeviceCreate:
      type: object
      required: [ip_address, hostname, traffic_gb]
      properties:
        ip_address:
          type: string
        hostname:
          type: string
        traffic_gb:
          type: number
        device_type:
          type: string
    SimulationDevice:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        device:
          $ref: '#/components/schemas/Device'
...